Update user unit tests for required password and optional phone number and change file structure.
